electrolyzer-digital-twin/software_stack/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
influxdb-client==1.38.0
requests==2.31.0
python-multipart==0.0.6
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0  # للـ AI (Isolation Forest)
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0  # لقراءة .env
PyJWT==2.8.0  # لـ API auth (بسيط)